+++ src/main/java/nz/co/mirality/refinedstorage4computercraft/data/LuaDoc.java
@@ -1,0 +1,13 @@
+package nz.co.mirality.refinedstorage4computercraft.data;
+
+import java.lang.annotation.*;
+
+@Documented
+@Retention(RetentionPolicy.RUNTIME)
+@Target({ElementType.METHOD})
+public @interface LuaDoc {
+    String args() default "";
+    String returns() default "";
+    int group() default Integer.MAX_VALUE;
+    int order() default Integer.MAX_VALUE;
+}