+++ src/main/resources/data/computercraft/lua/rom/autorun/refinedstorage.lua
@@ -1,0 +1,12 @@
+_G.refinedstorage = peripheral.find("refinedstorage")
+
+if refinedstorage then
+    local completion = require "cc.shell.completion"
+
+    shell.setPath(shell.path() .. ":/rom/programs/refinedstorage")
+    help.setPath(help.path() .. ":/rom/help/refinedstorage")
+
+    shell.setCompletionFunction("rom/programs/refinedstorage/stock.lua", completion.build(
+        { completion.choice, { "list", "item ", "fluid ", "delete ", "refresh", "refresh ", "start" } }
+    ))
+end